Pass a top-level navigation initiator origin to Fetch
To un-logjam the cookie layering work, I've started whatwg/fetch#1807.
That depends on this info to be piped into Fetch so we can
actually specify in WHATWG what SameSite=Strict means.

This patch plumbs that through on top-level navigatable fetches.

This doesn't build because it relies upon the corresponding
patch in Fetch. Let me know to land these.